% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCorCoeffs.R
\name{compareCorCoeffs}
\alias{compareCorCoeffs}
\title{This function returns p-values assessing the difference between two sets of correlation coefficients based on their bootstrapped confidence intervals}
\usage{
compareCorCoeffs(corrng1, corrng2)
}
\arguments{
\item{corrng1}{Table 1 with annotations and correlation coefficients}

\item{corrng2}{Table 2 with annotations and correlation coefficients}
}
\value{
Matrix with p-values
}
\description{
This function returns p-values assessing the difference between two sets of correlation coefficients based on their bootstrapped confidence intervals
}
