% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCorDistributions.R
\name{compareCorDistributions}
\alias{compareCorDistributions}
\title{This function returns p-values assessing the difference between two bootstrap distributions of correlation coefficients based on the Kolmogorov-Smirnov test}
\usage{
compareCorDistributions(corrng1, corrng2)
}
\arguments{
\item{corrng1}{Table 1 with annotations and correlation coefficients}

\item{corrng2}{Table 2 with annotations and correlation coefficients}
}
\value{
Matrix with p-values
}
\description{
This function returns p-values assessing the difference between two bootstrap distributions of correlation coefficients based on the Kolmogorov-Smirnov test
}
