% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReads.R
\name{countReads}
\alias{countReads}
\title{This function counts reads in a GRanges object that fall into a list of ROIs}
\usage{
countReads(data, rois, mino = 15, ignore_strand = TRUE, chr = 0)
}
\arguments{
\item{data}{GRanges object containing reads}

\item{rois}{GRanges object containing ROIs}

\item{mino}{Minimum overlap (see package 'GenomicRanges')}

\item{ignore_strand}{When set to TRUE, the strand information is ignored in the overlap calculations (see package 'GenomicRanges')}

\item{chr}{Chromosome number. When set to 0, all chromosomes will be considered.}
}
\value{
GRanges object containing read counts in ROIs
}
\description{
This function counts reads in a GRanges object that fall into a list of ROIs
}
