% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropAccMatrix.R
\name{cropAccMatrix}
\alias{cropAccMatrix}
\title{This function crops accessibility matrices}
\usage{
cropAccMatrix(am, start, end)
}
\arguments{
\item{am}{Accessibility matrix}

\item{start}{Start position of genomic region of interest}

\item{end}{End position of genomic region of interest}
}
\value{
Accessibility matrix without empty rows
}
\description{
This function crops accessibility matrices
}
