% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropCorMatrix.R
\name{cropCorMatrix}
\alias{cropCorMatrix}
\title{This function crops correlation matrices}
\usage{
cropCorMatrix(cm, chr, start, end)
}
\arguments{
\item{cm}{Correlation matrix}

\item{chr}{Chromosome for genomic region of interest}

\item{start}{Start position of genomic region of interest}

\item{end}{End position of genomic region of interest}
}
\value{
Accessibility matrix without empty rows
}
\description{
This function crops correlation matrices
}
