% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccLUT.R
\name{getAccLUT}
\alias{getAccLUT}
\title{This function makes a color LUT for accessibility matrices}
\usage{
getAccLUT(steps = 255)
}
\arguments{
\item{steps}{Number of steps for the color LUT}
}
\value{
Color LUT
}
\description{
This function makes a color LUT for accessibility matrices
}
