% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCIs.R
\name{getCIs}
\alias{getCIs}
\title{This function returns confidence intervals}
\usage{
getCIs(corrng, confidence = 0.95)
}
\arguments{
\item{corrng}{Table with annotations and correlation coefficients}

\item{confidence}{Confidence level (between 0 and 1)}
}
\value{
Stack of matrices containing correlation coefficients (first layer) along with minimum, maximum and median values from bootstrapping (second, third and fourth layer)
}
\description{
This function returns confidence intervals
}
