% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorBackground.R
\name{getCorBackground}
\alias{getCorBackground}
\title{This function returns the background correlation obtained with permuted entries}
\usage{
getCorBackground(am1, am2)
}
\arguments{
\item{am1}{Accessibility matrix 1}

\item{am2}{Accessibility matrix 2}
}
\value{
Matrix with background correlations (quantiles)
}
\description{
This function returns the background correlation obtained with permuted entries
}
