% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorLUT.R
\name{getCorLUT}
\alias{getCorLUT}
\title{This function makes a color LUT for correlation matrices}
\usage{
getCorLUT(steps = 255)
}
\arguments{
\item{steps}{Number of steps for the color LUT}
}
\value{
Color LUT
}
\description{
This function makes a color LUT for correlation matrices
}
