% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorRanges.R
\name{getCorRanges}
\alias{getCorRanges}
\title{This function returns correlation coefficients and their confidence intervals}
\usage{
getCorRanges(accmat, replicates = 100, confidence = 0.95)
}
\arguments{
\item{accmat}{Accessibility matrix}

\item{replicates}{Number of replicates for the bootstrap}

\item{confidence}{Confidence level (between 0 and 1)}
}
\value{
Stack of matrices containing correlation coefficients (first layer) along with minimum, maximum and median values from bootstrapping (second, third and fourth layer)
}
\description{
This function returns correlation coefficients and their confidence intervals
}
