% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPEPairs.R
\name{getPEPairs}
\alias{getPEPairs}
\title{This function returns promoter-enhancer (P-E) pairs based on a correlation matrix}
\usage{
getPEPairs(cm, cutoff, window = 0)
}
\arguments{
\item{cm}{Correlation matrix (rows: promoters, columns: enhancers)}

\item{cutoff}{Cutoff value required for P-E assignment}

\item{window}{Genomic window for P-E assignment}
}
\value{
P-E pairs
}
\description{
This function returns promoter-enhancer (P-E) pairs based on a correlation matrix
}
