% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCorMatrix.R
\name{makeCorMatrix}
\alias{makeCorMatrix}
\title{This function makes correlation matrices from accessibility matrices}
\usage{
makeCorMatrix(accmat, nmax = 0)
}
\arguments{
\item{accmat}{Accessibility matrix}

\item{nmax}{Number of cells used for the analysis. When set to 0, all cells will be considered.}
}
\value{
Correlation matrix
}
\description{
This function makes correlation matrices from accessibility matrices
}
