% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCrossCorMatrix.R
\name{makeCrossCorMatrix}
\alias{makeCrossCorMatrix}
\title{This function makes cross correlation matrices from different sets of regions within accessibility matrices}
\usage{
makeCrossCorMatrix(accmat1, accmat2, nmax = 0)
}
\arguments{
\item{accmat1}{Accessibility matrix 1 (data frame)}

\item{accmat2}{Accessibility matrix 2 (data frame)}

\item{nmax}{Number of cells used for the analysis. When set to 0, all cells will be considered.}
}
\value{
Cross correlation matrix
}
\description{
This function makes cross correlation matrices from different sets of regions within accessibility matrices
}
