% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGenomeTiles.R
\name{makeGenomeTiles}
\alias{makeGenomeTiles}
\title{This function makes tiles for the complete hg19 genome}
\usage{
makeGenomeTiles(tile_size)
}
\arguments{
\item{tile_size}{Tile size}
}
\value{
GRanges object with tiles for the complete genome
}
\description{
This function makes tiles for the complete hg19 genome
}
