% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIPCHist.R
\name{makeIPCHist}
\alias{makeIPCHist}
\title{This function makes a histogram for the number of integrations per cell}
\usage{
makeIPCHist(am)
}
\arguments{
\item{am}{Accessibility matrix}
}
\value{
Histogram (plot object)
}
\description{
This function makes a histogram for the number of integrations per cell
}
