% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIPCStats.R
\name{makeIPCStats}
\alias{makeIPCStats}
\title{Calculate statistics for the number of integrations per cell}
\usage{
makeIPCStats(am)
}
\arguments{
\item{am}{Accessibility matrix}
}
\description{
Calculate statistics for the number of integrations per cell
}
