% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIPRStats.R
\name{makeIPRStats}
\alias{makeIPRStats}
\title{Calculate statistics for the number of integrations per ROI}
\usage{
makeIPRStats(matrix)
}
\arguments{
\item{matrix}{Accessibility matrix}
}
\description{
Calculate statistics for the number of integrations per ROI
}
