% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTiles.R
\name{makeTiles}
\alias{makeTiles}
\title{This function makes tiles for a particular chromosome of the hg19 genome}
\usage{
makeTiles(chr = 1, start = 0, end = 0, tile_size = 1000,
  ignore_strand = FALSE)
}
\arguments{
\item{chr}{Chromosome number}

\item{start}{Start position (if only tiles for a particular region should be made)}

\item{end}{End position (if only tiles for a particular region should be made)}

\item{tile_size}{Tile size}

\item{ignore_strand}{When set to TRUE, the strand information is ignored in the overlap calculations (see package 'GenomicRanges')}
}
\value{
GRanges object with tiles for the indicated chromosome
}
\description{
This function makes tiles for a particular chromosome of the hg19 genome
}
