% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPairs.R
\docType{class}
\name{PEPairs-class}
\alias{PEPairs-class}
\title{PEPairs class that is used to store sets of promoter-enhancer pairs}
\description{
PEPairs class that is used to store sets of promoter-enhancer pairs
}
