% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBed.R
\name{readBed}
\alias{readBed}
\title{This function reads a BED file and stores the reads in a GRanges object}
\usage{
readBed(filename, chr = 1, ignore_strand = FALSE)
}
\arguments{
\item{filename}{File name of BED file}

\item{chr}{Chromosome number}

\item{ignore_strand}{When set to TRUE, the strand information is ignored in the overlap calculations (see package 'GenomicRanges')}
}
\value{
GRanges object with reads
}
\description{
This function reads a BED file and stores the reads in a GRanges object
}
