% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEmptyRows.R
\name{removeEmptyRows}
\alias{removeEmptyRows}
\title{This function removes empty rows from accessibility matrices}
\usage{
removeEmptyRows(am)
}
\arguments{
\item{am}{Accessibility matrix}
}
\value{
Accessibility matrix without empty rows
}
\description{
This function removes empty rows from accessibility matrices
}
