% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeRowsBelowCutoff.R
\name{removeRowsBelowCutoff}
\alias{removeRowsBelowCutoff}
\title{This function removes rows below a certain cutoff from accessibility matrices}
\usage{
removeRowsBelowCutoff(am, cutoff)
}
\arguments{
\item{am}{Accessibility matrix}

\item{cutoff}{Percentage of cells required to keep a row (should be between 0-100)}
}
\value{
Accessibility matrix without rows below a certain cutoff
}
\description{
This function removes rows below a certain cutoff from accessibility matrices
}
