% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveImage.R
\name{saveImage}
\alias{saveImage}
\title{This function saves a matrix as an image}
\usage{
saveImage(mat, path, w, h, color, min = 0, max = -1)
}
\arguments{
\item{mat}{Matrix}

\item{path}{Storage path for image}

\item{w}{Image width}

\item{h}{Image height}

\item{color}{Color LUT to be used}

\item{min}{Minimum value for color LUT}

\item{max}{Maximum value for color LUT}
}
\description{
This function saves a matrix as an image
}
