% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleCorMatrix.R
\name{scaleCorMatrix}
\alias{scaleCorMatrix}
\title{This function scales correlation matrices to genomic coordinates}
\usage{
scaleCorMatrix(cormat, size = 1000, start = 0, end = 249250621)
}
\arguments{
\item{cormat}{Correlation matrix (data frame)}

\item{size}{Number of genomic regions to be used}

\item{start}{Minimum genomic coordinate}

\item{end}{Maximum genomic coordinate}
}
\value{
Scaled correlation matrix
}
\description{
This function scales correlation matrices to genomic coordinates
}
