% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMatrix.R
\name{writeMatrix}
\alias{writeMatrix}
\title{Wrapper for making and storing scaled corrleation matrices within 1MB around a defined region}
\usage{
writeMatrix(am, chr, anno_start, anno_end, vcutoff, size, path)
}
\arguments{
\item{am}{Accessibility matrix}

\item{chr}{Chromosome for region of interest}

\item{anno_start}{Start of region of interest}

\item{anno_end}{End of region of interest}

\item{vcutoff}{Use only regions that are visible in at least this percentage of cells}

\item{path}{Storage path for scaled correlation matrix}
}
\description{
Wrapper for making and storing scaled corrleation matrices within 1MB around a defined region
}
