% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorRanges.R
\name{getCorRanges}
\alias{getCorRanges}
\title{This function returns bootstrapped correlation coefficients}
\usage{
getCorRanges(accmat, replicates = 100)
}
\arguments{
\item{accmat}{Accessibility matrix}

\item{replicates}{Number of replicates for the bootstrap}
}
\value{
Table with correlation coefficients
}
\description{
This function returns bootstrapped correlation coefficients
}
