% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCrossCorMatrix.R
\name{makeCrossCorMatrix}
\alias{makeCrossCorMatrix}
\title{This function makes a cross correlation matrix from two accessibility matrices}
\usage{
makeCrossCorMatrix(am1, am2, nmax = 0)
}
\arguments{
\item{am1}{Accessibility matrix 1}

\item{am2}{Accessibility matrix 2}

\item{nmax}{Number of cells used for the analysis. When set to 0, all cells will be considered.}
}
\value{
Cross correlation matrix
}
\description{
This function makes a cross correlation matrix from two accessibility matrices
}
