% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIPRHist.R
\name{makeIPRHist}
\alias{makeIPRHist}
\title{This function makes a histogram for the number of integrations per ROI}
\usage{
makeIPRHist(matrix)
}
\arguments{
\item{matrix}{Accessibility matrix}
}
\value{
Histogram (as a plot object)
}
\description{
This function makes a histogram for the number of integrations per ROI
}
