% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleAccMatrix.R
\name{scaleAccMatrix}
\alias{scaleAccMatrix}
\title{This function scales accessibility matrices to genomic coordinates}
\usage{
scaleAccMatrix(am, chr, size = 1000, start = 0, end = 249250621)
}
\arguments{
\item{am}{Accessibility matrix}

\item{chr}{Chromosome to be used}

\item{size}{Number of genomic regions to be used}

\item{start}{Minimum genomic coordinate}

\item{end}{Maximum genomic coordinate}
}
\value{
Scaled accessibility matrix
}
\description{
This function scales accessibility matrices to genomic coordinates
}
