% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMatrix.R
\name{writeMatrix}
\alias{writeMatrix}
\title{Wrapper for making and storing scaled corrleation matrices within 1MB around a defined region}
\usage{
writeMatrix(accmat, anno_start, anno_end, cutoff, path)
}
\arguments{
\item{accmat}{Accessibility matrix}

\item{anno_start}{Start of region of interest}

\item{anno_end}{End of region of interest}

\item{cutoff}{Use only regions that are visible in at least this number of cells.}

\item{path}{Storage path for scaled correlation matrix}
}
\description{
Wrapper for making and storing scaled corrleation matrices within 1MB around a defined region
}
